+++ examples/eqAsync/main.cpp
@@ -1,6 +1,6 @@
 
-/* Copyright (c) 2007-2014, Stefan Eilemann <eile@equalizergraphics.com>
- *               2010-2011, Maxim Makhinya  <maxmah@gmail.com>
+/* Copyright (c) 2007-20147, Stefan Eilemann <eile@equalizergraphics.com>
+ *                           Maxim Makhinya  <maxmah@gmail.com>
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -48,6 +48,17 @@ public:
 
 int main( const int argc, char** argv )
 {
+    for( int i=1; i < argc; ++i )
+    {
+        if( std::string( argv[ i ]) == "--help" )
+        {
+            std::cout << lunchbox::getFilename( argv[0] )
+                      << ": asynchronous GPU upload example" << std::endl
+                      << eq::getHelp() << eq::Client::getHelp() << std::endl;
+            return EXIT_SUCCESS;
+        }
+    }
+
     // 1. Equalizer initialization
     NodeFactory nodeFactory;
     if( !eq::init( argc, argv, &nodeFactory ))